♻️ 홈 화면 동물 카드 슬라이드 AnimalSlide로 분리
